createTodo and getTodo are working correctly
I wanted to make it so that there can't be two
of the same exactly named todos in a project
at the same time, so I added a uniqueness
check during todo creation. I also added
a function that searches for a specific todo
given a project name and a todo title name.
Since you can only create todos that are unique
inside that project, the check is not needed in
getTodo