Added removeTodo, refactored entire project.js
So as I worked on project.js and todo.js, I
noticed that I had to import project.js into
todo.js in order to interact with it since
todos are inside the project object. This
created a lot of dependencies between the modules
and overtime I realized that it didn't make much
sense to separate todo and project into two
different modules since they are apart of the
same object/data. If module has to use/modify the
project object, then it would violate the
single responsibility principle. So I refactored
it so that both modules are now one module that
controls our main localStorage data, both
for projects and todos inside the projects.