Added connection to TodayQuest.js to app logic
Dynamically displays either the first
project from localStorage or a newly made
default project if localStorage is empty.
First, gets the project object from localStorage
then passes that project object to our
generateProjHeader function to create the
header with the current project's name on it, and
also passes it in a loop to the generateTodoCard
to display all of the todos in the todos array.