Added blank proj name check, alerts if name exists
To ensure that the user will put in a valid,
non-empty name, I added a check that
checks if the FormData is an empty string.
If it is, then it will output an alert message
letting the user know that the name is invalid.
Also included a check if the name already
exists in localStorage, then it alerts that
the project name already exists.