Added boolean returns for updateProjName
To ensure that the modal doesn't close and the
page doesn't re-render, I added an if statement
check to see if the name already exists
in the localStorage upon clicking submit.
If the name does exist, then it will cause the
if statement to prevent the modal and page
from re-loading. If it doesn't exist, then
proceed with re-loading the page. Updates the
project name correctly in localStorage.