Finished change quest button logic, fixed update
There was an issue with updateTodo where the
function wasn't checking if the previous
todo to be replaced even existed, so I added
a check for that first, then it will attempt
to replace it. Also finished logic for the
project switching. My main issue was that the
tempProj variable in the generateTodayQuest()
was unsharable, and since the entire page
rendered on it, I needed it to change somehow,
so I instead converted it into a parameter
default value. If the user doesn't have
any projects, then it will still make one.
If it does, then it will display the
last project in localStorage which could be
anything. This way, I can pass any current
tempProject into the re-rendering process
to stay on the same page or switch pages
if needed.