Finished view button modal functionality
For each todo, the code dynamically generates a
todo view modal for itself. It only updates on
re-renders/re-loads of the entire homepage, but
it displays the chosen todo's information fully
with the selected stylings. I had an issue where
the dialog wouldn't hide itself and I discovered
that it was a CSS issue, where I accidentally
set the modal/dialog to display: flex;, which
makes it not hide. I also needed a way to
make each modal unique from one another, so I
set it to make a class tag for each modal with
a script `${todoObj.title.replaceAll(" ", "-")}`
where each todo modal will have its own unique
name with spaces filled with dashes so that
it can call it's own unique modal.