Working on todo edit button modal
Currently constructed the todo edit button modal
and all of it's buttons/Inputs for the form.
I am coding the submit button at the moment and
creating checks to see if the data entered is an
empty string or an empty space, in which the
form/modal will not submit or re-render and will
send an alert to the user if it is an invalid
input.