Fixed re-render on edit todo button
The main issue I had was it was re-rendering
the currentProj object which was the object
passed to us BEFORE the change happened, so I
just told the appManage to get the project again
AFTER the change. Confirmed this is working
5/6/2025, and will use the same method to fix the
same re-rendering issue with the change project
name button.