fixed re-rendering but with change project name
When changing the project name, the code had an
issue with the today's quest page re-rendering
the wrong project because of the name change. I
added another appManage.getProject(); to search
specifically for the modalData.get("projTitle"),
which is the user-given project title AFTER the
change. So the sequence goes like this:

re-write current project's name -> get project
with NEW name from form input -> re-render page
with the project but with the NEW name.