Fixed same todo name error when editing todo
When pressing the edit todo button, the fields
in the form get filled with the already
existing content of the todo so that the
user can optionally edit if needed. The
program prevented two of the same todo, but
since that only matters for projects,
I deleted the code that checks if there are the
same todo, since you are allowed to make
multiple todos under the same project with the
same name.