Finished the page switching logic successfully
One issue I had was that the page kept refreshing
everytime I clicked one of the navigation
links to the left side, but that was because the
anchor element by default, refreshes/redirects
the link to somewhere, so I added a preventDefault
to ensure that the links do not refresh the
page. Now everything works as intended. Everytime
a user clicks on a link, it will re-render
the new page. Each page is functional. Will add
some tweaks and fix bugs if needed.