Updated README, finished project
I am declaring this project finished and Updated
the README file to contain my final thoughts and
updates on what is working and when I finished it.
This project was complex, at least in my eyes.
It let me demonstrate all of the principles
and things I learned in the previous lessons.
It also let me work with JSON files! This was
an awesome beginner project which I recommend
anyone to try out for themselves.